+++ packages/example/src/components/connector/environments.ts
@@ -1,10 +1,7 @@
 export const ENVIRONMENTS = [{
 	value: "prod",
 	label: "Production"
-},/* {
-	value: "dev",
-	label: "Dev"
-}, */{
+}, {
 	value: "development",
 	label: "Development"
 }, {