+++ packages/sdk-examples/src/backend/tezos/accept-bid.ts
@@ -13,7 +13,7 @@ async function acceptBid() {
 		in_memory_provider(
 			"edskRqrEPcFetuV7xDMMFXHLMPbsTawXZjH9yrEz4RBqH1" +
       "D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj",
-			"https://dev-tezos-node.rarible.org"
+			"https://rpc.tzkt.io/ithacanet"
 		)
 	)
 	const sdk = createRaribleSdk(wallet, "testnet")