+++ packages/sdk/src/sdk-blockchains/ethereum/fill.ts
@@ -104,6 +104,14 @@ export class EthereumFill {
 				}
 				break
 			}
+			case "LOOKSRARE": {
+				request = {
+					order,
+					originFees: toEthereumParts(fillRequest.originFees),
+					amount: fillRequest.amount,
+				}
+				break
+			}
 			default: {
 				throw new Error("Unsupported order type")
 			}
@@ -174,6 +182,14 @@ export class EthereumFill {
 					payoutsSupport: PayoutsSupport.NONE,
 					maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
 					supportsPartialFill,
+				}
+			}
+			case "LOOKSRARE": {
+				return {
+					originFeeSupport: OriginFeeSupport.FULL,
+					payoutsSupport: PayoutsSupport.NONE,
+					maxFeesBasePointSupport: MaxFeesBasePointSupport.IGNORED,
+					supportsPartialFill: true,
 				}
 			}
 			default: